add deleteMarkers method|add dark mode style to map after 6PM|add user data banner to mapActivity

+++ app/src/main/java/com/example/myapp/Model/MarkerModel.kt
@@ -21,5 +21,5 @@ data class MarkerModel(
     var lat: Double? = null,
     var long: Double? = null,
     var top_score: Int? = null,
-    var time_to_live: Int?=null
+    var time_to_live: Long?=null
 )
